app/requirements.txt
Django
python-socketio
eventlet
redis
celery
djangorestframework
drf-yasg
django-braces
python-dotenv